fix(publish): location of the compiled results
it's not really necessary to stay "backwards compatible" here, because older versions of LNReader never made assumptions about these paths to begin with. And it only causes issues with the local test hosting. So for now I've made it so that both paths are ACTUALLY available, but only the new one is used is possible.